requirements.txt
streamlit
pymongo
python-dotenv
bcrypt
python-docx
